Move `core/convar/` to `tier1/`
