+++ primedev/Northstar.cmake
@@ -26,12 +26,6 @@ add_library(
     "client/rejectconnectionfixes.cpp"
     "config/profile.cpp"
     "config/profile.h"
-    "core/convar/concommand.cpp"
-    "core/convar/concommand.h"
-    "core/convar/convar.cpp"
-    "core/convar/convar.h"
-    "core/convar/cvar.cpp"
-    "core/convar/cvar.h"
     "core/filesystem/filesystem.cpp"
     "core/filesystem/filesystem.h"
     "core/filesystem/rpakfilesystem.cpp"
@@ -139,6 +133,12 @@ add_library(
     "squirrel/squirrelautobind.h"
     "squirrel/squirrelclasstypes.h"
     "squirrel/squirreldatatypes.h"
+    "tier1/cmd.cpp"
+    "tier1/cmd.h"
+    "tier1/convar.cpp"
+    "tier1/convar.h"
+    "tier1/cvar.cpp"
+    "tier1/cvar.h"
     "util/printcommands.cpp"
     "util/printcommands.h"
     "util/printmaps.cpp"
