+++ primedev/mods/modmanager.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "core/convar/convar.h"
+#include "tier1/convar.h"
 #include "core/memalloc.h"
 #include "squirrel/squirrel.h"
