+++ primedev/plugins/plugins.cpp
@@ -4,7 +4,7 @@
 #include "squirrel/squirrel.h"
 #include "plugins.h"
 #include "masterserver/masterserver.h"
-#include "core/convar/convar.h"
+#include "tier1/convar.h"
 #include "server/serverpresence.h"
 #include <optional>
 #include <regex>
