+++ primedev/server/serverpresence.cpp
@@ -1,7 +1,7 @@
 #include "serverpresence.h"
 #include "shared/playlist.h"
 #include "core/tier0.h"
-#include "core/convar/convar.h"
+#include "tier1/convar.h"
 
 #include <regex>
