+++ primedev/shared/exploit_fixes/ns_limits.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "engine/r2engine.h"
-#include "core/convar/convar.h"
+#include "tier1/convar.h"
 #include <unordered_map>
 
 struct PlayerLimitData
