+++ primedev/tier1/cmd.cpp
@@ -1,4 +1,4 @@
-#include "concommand.h"
+#include "tier1/cmd.h"
 #include "shared/misccommands.h"
 #include "engine/r2engine.h"
