+++ primedev/tier1/convar.h
@@ -2,7 +2,7 @@
 #include "core/sourceinterface.h"
 #include "core/math/color.h"
 #include "cvar.h"
-#include "concommand.h"
+#include "tier1/cmd.h"
 
 // taken directly from iconvar.h
