+++ primedev/util/printcommands.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "core/convar/concommand.h"
+#include "tier1/cmd.h"
 
 void PrintCommandHelpDialogue(const ConCommandBase* command, const char* name);
 void TryPrintCvarHelpForCommand(const char* pCommand);
